+++ TrayX/App.xaml
@@ -13,6 +13,9 @@
                 <ContextMenu>
                     <MenuItem Header="🖥️ Show Window" Click="Tray_ShowWindow"/>
                     <MenuItem Header="🧹 Clean RAM" Click="Tray_CleanRam"/>
+                    <MenuItem Header="🧹 Clear Temp" Click="Tray_ClearTemp"/>
+                    <MenuItem Header="🌐 Flush DNS" Click="Tray_FlushDns"/>
+                    <MenuItem Header="🗑 Empty Recycle Bin" Click="Tray_EmptyRecycleBin"/>
                     <Separator/>
                     <MenuItem Header="❌ Exit" Click="Tray_Exit"/>
                 </ContextMenu>
