+++ TrayX/MainWindow.xaml
@@ -153,8 +153,10 @@
                 </Grid>
 
                 <StackPanel Margin="0 10 0 10">
-                    <lvc:CartesianChart Series="{Binding CpuSeries}" Height="80"/>
-                    <lvc:CartesianChart Series="{Binding RamSeries}" Height="80" Margin="0,5,0,0"/>
+                    <TextBlock Text="CPU Usage (last 60s)" FontSize="12" Foreground="#BBBBBB"/>
+                    <lvc:CartesianChart Series="{Binding CpuSeries}" Height="80" Margin="0,0,0,5"/>
+                    <TextBlock Text="RAM Usage (last 60s)" FontSize="12" Foreground="#BBBBBB"/>
+                    <lvc:CartesianChart Series="{Binding RamSeries}" Height="80"/>
                 </StackPanel>
 
                 <!-- Disks -->
@@ -199,21 +201,6 @@
                                TextWrapping="Wrap"/>
                 </Border>
 
-                <!-- Maintenance -->
-                <TextBlock Text="🧰 Maintenance"
-                           FontSize="13" FontWeight="Bold"
-                           Foreground="#BBBBBB" Margin="0 0 0 5"/>
-                <UniformGrid Columns="3" Margin="0,0,0,10">
-                    <Button Style="{StaticResource CleanButtonStyle}"
-                            Content="🧹 Clear Temp"
-                            Click="ClearTemp_Click"/>
-                    <Button Style="{StaticResource CleanButtonStyle}"
-                            Content="🌐 Flush DNS"
-                            Click="FlushDns_Click"/>
-                    <Button Style="{StaticResource CleanButtonStyle}"
-                            Content="🗑 Empty Recycle Bin"
-                            Click="EmptyRecycleBin_Click"/>
-                </UniformGrid>
 
             </StackPanel>
         </ScrollViewer>
