+++ TrayX/MainWindow.xaml.cs
@@ -253,85 +253,5 @@ public partial class MainWindow
         return val >= 1 ? $"{val:0.00} Mbps" : $"{val * 1000:0} Kbps";
     }
     
-    private void ClearTemp_Click(object sender, RoutedEventArgs e)
-    {
-        try
-        {
-            var tempPath = Environment.GetEnvironmentVariable("TEMP");
-            if (string.IsNullOrEmpty(tempPath)) return;
-            Process.Start(new ProcessStartInfo
-            {
-                FileName = "cmd.exe",
-                Arguments = $"/C del /q/f/s \"{tempPath}\\*\"",
-                WindowStyle = ProcessWindowStyle.Hidden,
-                CreateNoWindow = true
-            });
-
-            MessageBox.Show("Temporary files deleted (or scheduled for deletion).", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
-        }
-        catch (Exception ex)
-        {
-            ErrorLogger.LogException(ex);
-            MessageBox.Show("An error occurred while clearing temporary files. Please check the log for details.",
-                "Error", MessageBoxButton.OK, MessageBoxImage.Error);
-        }
-    }
-
-    private void FlushDns_Click(object sender, RoutedEventArgs e)
-    {
-        try
-        {
-            Process.Start(new ProcessStartInfo
-            {
-                FileName = "ipconfig",
-                Arguments = "/flushdns",
-                WindowStyle = ProcessWindowStyle.Hidden,
-                CreateNoWindow = true
-            });
-
-            MessageBox.Show("DNS cache flushed successfully.", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
-        }
-        catch (Exception ex)
-        {
-            ErrorLogger.LogException(ex);
-            MessageBox.Show("An error occurred while flushing the DNS cache. Please check the log for details.",
-                "Error", MessageBoxButton.OK, MessageBoxImage.Error);
-        }
-    }
-    
-    private void EmptyRecycleBin_Click(object sender, RoutedEventArgs e)
-    {
-        try
-        {
-            var flags = NativeMethods.SherbNoconfirmation | NativeMethods.SherbNoprogressui | NativeMethods.SherbNosound;
-            var result = NativeMethods.SHEmptyRecycleBin(IntPtr.Zero, null, flags);
-
-            if (result == 0)
-            {
-                MessageBox.Show("Recycle Bin emptied successfully.", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
-            }
-            else
-            {
-                MessageBox.Show($"Failed to empty Recycle Bin. Error code: {result}", "Error", MessageBoxButton.OK, MessageBoxImage.Warning);
-            }
-        }
-        catch (Exception ex)
-        {
-            ErrorLogger.LogException(ex);
-            MessageBox.Show("An error occurred while emptying the Recycle Bin. Please check the log for details.",
-                "Error", MessageBoxButton.OK, MessageBoxImage.Error);
-        }
-    }
-
-
 }
-}
-internal static class NativeMethods
-{
-    [DllImport("shell32.dll", SetLastError = true)]
-    internal static extern uint SHEmptyRecycleBin(IntPtr hwnd, string? pszRootPath, uint dwFlags);
-
-    internal const uint SherbNoconfirmation = 0x00000001;
-    internal const uint SherbNoprogressui   = 0x00000002;
-    internal const uint SherbNosound        = 0x00000004;
 }
